\name{GetALetPack-package}
\alias{GetALetPack-package}
\alias{GetALetPack}
\docType{package}
\title{
\packageTitle{GetALetPack}
}
\description{
\packageDescription{GetALetPack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GetALetPack}
\packageIndices{GetALetPack}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GetALetPack}

Maintainer: \packageMaintainer{GetALetPack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
