\name{RandLetter-package}
\alias{RandLetter-package}
\alias{RandLetter}
\docType{package}
\title{
\packageTitle{RandLetter}
}
\description{ generating Random Letter
\packageDescription{RandLetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RandLetter}
\packageIndices{RandLetter}
It basically randomly generates a random letter from the 26 letter alphabet
}
\author{
\packageAuthor{RandLetter}

Maintainer:Joselyne Umubyeyi \packageMaintainer{RandLetter}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
